% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edererIIhaz.R
\name{edererIIhaz}
\alias{edererIIhaz}
\title{Ederer II hazard estimator}
\usage{
edererIIhaz(.data, .time, .event, .matchedhaz)
}
\arguments{
\item{.data}{data.frame}

\item{.time}{Integer. Survival time}

\item{.event}{0/1 or 'TRUE'/'FALSE'. Censor / Survival status indicator}

\item{estimate}{"surv" (point estimate), "lower" (lower conf limit), "upper" (upper conf limit)}
}
\value{
A tibble
}
\description{
Calculate mean expected risk/hazard/mortality using the Ederer II method (i.e., only for those who remain at risk)
}
\keyword{survival}
