% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmstepfun.R
\name{kmest}
\alias{kmest}
\title{KM estimator}
\usage{
kmest(day, time, event, estimate = c("surv", "lower", "upper"))
}
\arguments{
\item{day}{Integer. Number of days after time zero}

\item{time}{Integer. Survival time}

\item{event}{0/1 or 'TRUE'/'FALSE'. Censor / Survival status indicator}

\item{estimate}{"surv" (point estimate), "lower" (lower conf limit), "upper" (upper conf limit)}
}
\value{
A function of class "stepfun"
}
\description{
Wrapper for [stepfun()] to creates a function to output vector of KM estimates for a given survival-censor pair
}
\keyword{kaplan-meier}
\keyword{survival,}
