% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rsurv.R
\name{tidy_rsurv}
\alias{tidy_rsurv}
\title{To tidy survfit object from [relsurv::rs.surv()]}
\usage{
tidy_rsurv(rsurvfit)
}
\arguments{
\item{rsurvfit}{Object of class "survfit" from [relsurv::rs.surv()]}
}
\value{
A tibble.
}
\description{
Adds various useful survival-based point estimates for survfit object
}
\keyword{relative}
\keyword{survival}
\keyword{survival,}
