% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_surv.R
\name{tidy_surv}
\alias{tidy_surv}
\title{To tidy up survfit object}
\usage{
tidy_surv(survfit, expand = FALSE)
}
\arguments{
\item{survfit}{Object of class "survfit" from [survival::Surv()]}

\item{expand}{Logical. if true then expand survival estimates to include rows for days when no events (or censoring) occurred.}
}
\value{
A tibble.
}
\description{
Add various other survival-based point estimates to a tidied survfit object. Like [broom::tidy()] but adds more estimates. Includes ability to expland o include all possible time points, even where no events occurred
}
\keyword{survival}
